# docs/swagger.yaml
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: ticket service
  license:
    name: MIT
    url: http://175.27.214.248/gitlab-instance-80f356dd/ticket/-/blob/main/LICENSE
  termsOfService: http://175.27.214.248:8888
  title: ticket API
  version: "1.0"
paths:
  /api/ticket/upload:
    post:
      parameters:
      - description: Image File
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Import image
      tags:
      - util
...
swagger: "2.0"